initial auth version
