+++ core/v1/auth/config.py
@@ -1,0 +1,3 @@
+SECRET_KEY = "09d25e094faa6ca2556c818166b7a9563b03f7099f6f0f4caa6cf63b88e8d3e7"
+ALGORITHM = "HS256"
+ACCESS_TOKEN_EXPIRE_MINUTES = 30
