+++ core/v1/auth/util.py
@@ -1,0 +1,37 @@
+from datetime import datetime, timedelta, timezone
+
+from passlib.context import CryptContext
+from jose import jwt
+
+from core.v1.auth.config import ALGORITHM
+from core.v1.auth.config import SECRET_KEY
+from core.v1.auth.model import TokenData
+
+
+pwd_context = CryptContext(schemes=["bcrypt"], deprecated="auto")
+
+
+def verify_password(plain_password, hashed_password):
+    return pwd_context.verify(plain_password, hashed_password)
+
+
+def get_password_hash(password):
+    return pwd_context.hash(password)
+
+
+def create_access_token(data: dict, expires_delta: timedelta | None = None):
+    to_encode = data.copy()
+    if expires_delta:
+        expire = datetime.now(timezone.utc) + expires_delta
+    else:
+        expire = datetime.now(timezone.utc) + timedelta(minutes=15)
+    to_encode.update({"exp": expire})
+    encoded_jwt = jwt.encode(to_encode, SECRET_KEY, algorithm=ALGORITHM)
+    return encoded_jwt
+
+
+def get_token_data(token: str):
+    payload = jwt.decode(token, SECRET_KEY, algorithms=[ALGORITHM])
+    username: str = payload.get("sub")
+    token_scopes = payload.get("scopes", [])
+    return TokenData(scopes=token_scopes, username=username)
