+++ core/v1/exc.py
@@ -1,0 +1,4 @@
+from fastapi import HTTPException
+
+
+class APIException(HTTPException): ...
